+++ src/server/graphql/blockchain/index.ts
@@ -1,10 +1,10 @@
 import { IResolvers } from "apollo-server";
 import { assignDeep } from "../../utils";
-import { resolvers as accountResolvers } from "./account";
+import { resolvers as accountResolvers } from "./types/account";
+import { resolvers as blockResolvers } from "./types/block";
+import { resolvers as messageResolvers } from "./types/message";
+import { resolvers as transactionResolvers } from "./types/transaction";
 import { resolvers as blockchainResolvers } from "./blockchain";
-import { resolvers as blockResolvers } from "./block";
-import { resolvers as messageResolvers } from "./message";
-import { resolvers as transactionResolvers } from "./transaction";
 
 const resolvers = {} as IResolvers;
 [