+++ src/server/graphql/blockchain/types/account.ts
@@ -1,9 +1,9 @@
 import { 
     BigIntArgs,
     resolveBigUInt,
-} from "../../filter/filters";
-import { QRequestContext } from "../../request";
-import { AccountStatusEnum, Resolvers } from "./resolvers-types-generated";
+} from "../../../filter/filters";
+import { QRequestContext } from "../../../request";
+import { AccountStatusEnum, Resolvers } from "../resolvers-types-generated";
 
 export const resolvers: Resolvers<QRequestContext> = {
     BlockchainAccount: {